Review Changes: Set only iconify props
